requirements.txt
python-dotenv
langchain
langchain-ollama
langchain-openai